refactor: refactoring some names and poms
