+++ pom.xml
@@ -15,11 +15,11 @@
 		<module>common</module>
 		<module>shortener-service</module>
 		<module>resolver-service</module>
-		<module>key-generator-service</module>
+		<module>keygen-service</module>
 	</modules>
 
 	<properties>
-		<java.version>17</java.version>
+		<java.version>21</java.version>
 		<spring-boot.version>3.4.5</spring-boot.version>
 	</properties>
 
@@ -41,7 +41,18 @@
 				<plugin>
 					<groupId>org.springframework.boot</groupId>
 					<artifactId>spring-boot-maven-plugin</artifactId>
+					<version>${spring-boot.version}</version>
 				</plugin>
+
+				<plugin>
+					<groupId>org.apache.maven.plugins</groupId>
+					<artifactId>maven-resources-plugin</artifactId>
+					<version>3.3.1</version>
+					<configuration>
+						<encoding>UTF-8</encoding>
+					</configuration>
+				</plugin>
+
 			</plugins>
 		</pluginManagement>
 	</build>
