+++ shortener-service/src/main/resources/application.yml
@@ -1,22 +1,22 @@
-server:
-  port: ${SERVER_PORT}
-
-spring:
-  application:
-    name: url-shortener
-  data:
-    mongodb:
-      host: ${MONGO_HOST}
-      port: ${MONGO_PORT}
-      uri: mongodb://${MONGO_HOST}:${MONGO_PORT}/${MONGO_DATABASE}
-      username: ${MONGO_USERNAME}
-      password: ${MONGO_PASSWORD}
-      database: ${MONGO_DATABASE}
-
-    redis:
-      host: ${REDIS_HOST}
-      port: ${REDIS_PORT}
-      password: ${REDIS_PASSWORD}
-      client-type: lettuce
-  cache:
-    type: ${SPRING_CACHE_TYPE}
+#server:
+#  port: ${SERVER_PORT}
+#
+#spring:
+#  application:
+#    name: url-shortener
+#  data:
+#    mongodb:
+#      host: ${MONGO_HOST}
+#      port: ${MONGO_PORT}
+#      uri: mongodb://${MONGO_HOST}:${MONGO_PORT}/${MONGO_DATABASE}
+#      username: ${MONGO_USERNAME}
+#      password: ${MONGO_PASSWORD}
+#      database: ${MONGO_DATABASE}
+#
+#    redis:
+#      host: ${REDIS_HOST}
+#      port: ${REDIS_PORT}
+#      password: ${REDIS_PASSWORD}
+#      client-type: lettuce
+#  cache:
+#    type: ${SPRING_CACHE_TYPE}
