+++ shortener-service/src/test/java/com/isaacandrade/urlshortenerservice/UrlShortenerApplicationTests.java
@@ -1,4 +1,4 @@
-package com.isaacandrade.url_shortener;
+package com.isaacandrade.urlshortenerservice;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
